# BED-4463: Implement BOM encoding package and refactor WriteAndValidateJSON for improved Unicode handling