--- a/cmd/api/src/services/fileupload/validation.go
+++ b/cmd/api/src/services/fileupload/validation.go
@@ -19,9 +19,10 @@ package fileupload
 import (
 	"encoding/json"
 	"errors"
+	"io"
+
 	"github.com/specterops/bloodhound/log"
 	"github.com/specterops/bloodhound/src/model/ingest"
-	"io"
 )
 
 var ZipMagicBytes = []byte{0x50, 0x4b, 0x03, 0x04}
@@ -52,7 +53,7 @@ func ValidateMetaTag(reader io.Reader, readToEnd bool) (ingest.Metadata, error) 
 				return ingest.Metadata{}, ErrInvalidJSON
 			}
 		} else {
-			//Validate that our data tag is actually opening correctly
+			// Validate that our data tag is actually opening correctly
 			if dataTagFound && !dataTagValidated {
 				if typed, ok := token.(json.Delim); ok && typed == ingest.DelimOpenSquareBracket {
 					dataTagValidated = true