--- a/go.work
+++ b/go.work
@@ -19,6 +19,7 @@ go 1.21
 use (
 	./cmd/api/src
 	./packages/go/analysis
+	./packages/go/bomenc
 	./packages/go/cache
 	./packages/go/conftool
 	./packages/go/crypto