… packages/go/bomenc/encodings_test.go
… packages/go/bomenc/go.sum
… packages/go/bomenc/normalize_test.go
… packages/go/bomenc/utf16.go
… packages/go/bomenc/utf16_test.go
… packages/go/bomenc/utf32.go
… packages/go/bomenc/utf32_test.go