<!-- README: https://github.com/SpecterOps/BloodHound/issues/672 -->
<!-- All pull requests require either an associated -->
<!-- Jira ticket or GitHub issue. PRs opened without -->
<!-- an associated discussion item will be closed! -->

## Description

This PR introduces a robust package for handling Byte Order Mark (BOM) encoding detection and UTF conversions and refactors the `WriteAndValidateJSON` function to use this new package. The changes include:

1. Implement an `Encoding` interface to handle different encodings.
2. Precise BOM detection for UTF-8, UTF-16 (BE/LE), and UTF-32 (BE/LE).
3. Implementation of UTF-16 and UTF-32 to UTF-8 conversion with proper handling of surrogate pairs.
4. Refactoring of `WriteAndValidateJSON` to use the new `bomenc` package for BOM handling and UTF normalization.
5. Improved error handling and simplified logic in `WriteAndValidateJSON`.

## Motivation and Context

This PR addresses: [BED-4463](https://specterops.atlassian.net/browse/BED-4463)

These changes are necessary to enhance our management of diverse text encodings across the codebase, especially in IO operations involving potentially BOM-prefixed content. The introduction of the `bomenc` package offers a more robust and comprehensive solution for handling various Unicode encodings. The refactoring of `WriteAndValidateJSON` leverages this package to simplify the function and enhance its versatility and reliability.

## How Has This Been Tested?

Developed comprehensive test suites for the `bomenc` package, including:

- Tests for encoding detection, conversion, and edge cases.
- Tests for custom encodings and the new `HasSequence` method.
- Tests covering various scenarios of BOM detection and removal.
- Tests for UTF normalization with different input encodings.

Testing the refactored `WriteAndValidateJSON` function to ensure correct behavior with inputs of various encodings and with/without BOMs.

## Types of changes

- New feature (non-breaking change which adds functionality)
- Chore (a change that does not modify the application functionality)

## Checklist:

- [x] I have met the contributing prerequisites
  - Assigned myself to this PR
  - Added the appropriate labels
  - Associated an issue: https://github.com/SpecterOps/BloodHound/issues/672
  - Read the Contributing guide: https://github.com/SpecterOps/BloodHound/wiki/Contributing
- [x] I have ensured that related documentation is up-to-date
  - Open API docs
  - Code comments (GoDocs / JSDocs)
- [x] I have followed proper test practices
  - Added/updated tests to cover my changes
  - All new and existing tests passed

[BED-4463]: https://specterops.atlassian.net/browse/BED-4463?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ